refactor: combine date and datetime fields
